Move System::stripRootDir() to the StringUtil class
