+++ src/Resources/contao/library/Contao/System.php
@@ -1185,30 +1185,4 @@ abstract class System
 	{
 		@trigger_error('Using System::disableModule() has been deprecated and will no longer work in Contao 5.0. Use Composer to add or remove modules.', E_USER_DEPRECATED);
 	}
-
-	/**
-	 * Strip the Contao root dir from the given absolute path
-	 *
-	 * @param string $path
-	 *
-	 * @return string
-	 *
-	 * @throws \InvalidArgumentException
-	 */
-	public static function stripRootDir($path)
-	{
-		static $length = null;
-
-		if ($length === null)
-		{
-			$length = strlen(TL_ROOT);
-		}
-
-		if (strncmp($path, TL_ROOT, $length) !== 0 || strlen($path) <= $length || ($path[$length] !== '/' && $path[$length] !== '\\'))
-		{
-			throw new \InvalidArgumentException(sprintf('Path is not inside the Contao root dir "%s"', $path));
-		}
-
-		return substr($path, $length + 1);
-	}
 }
