+++ tests/Fixtures/library/System.php
@@ -74,21 +74,4 @@ class System
     {
         $GLOBALS['TL_LANG'] = new LanguageHelper();
     }
-
-    public static function stripRootDir($path)
-    {
-        static $length = null;
-
-        if ($length === null)
-        {
-            $length = strlen(TL_ROOT);
-        }
-
-        if (strncmp($path, TL_ROOT, $length) !== 0 || strlen($path) <= $length || ($path[$length] !== '/' && $path[$length] !== '\\'))
-        {
-            throw new \InvalidArgumentException(sprintf('Path is not inside the Contao root dir "%s"', $path));
-        }
-
-        return substr($path, $length + 1);
-    }
 }
